pyproject.toml
[tool.poetry]
name = "soil-api"
version = "0.1.0"
description = "API to manage data for SOIL lab"
authors = ["Evan Thomas <evan.thomas@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlmodel = "^0.0.11"
geoalchemy2 = "^0.14.2"
aiosqlite = "^0.19.0"
httpx = "^0.25.1"
pytest-asyncio = "^0.21.1"
asyncpg = "^0.29.0"
shapely = "^2.0.2"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.12.1"
pydantic-geojson = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-watcher = "^0.3.4"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
